removing auth from api

+++ app/[locale]/admin/dashboard/clients/(fetch)/deleteClient.ts
@@ -4,43 +4,48 @@ import { getServerSession } from "next-auth";
 import { revalidatePath } from "next/cache";
 import { authOptions } from "@/app/models/db/authOptions";
 
-const EXPIRE_SECONDS = 15 * 24 * 60 * 60; // 15 days
-
 export async function deleteClient(clientId: string) {
   const session = await getServerSession(authOptions);
 
-  // ❌ Not logged in
   if (!session) {
-    throw new Error("UNAUTHENTICATED");
+    return {
+      message: "UNAUTHENTICATED",
+      success: false,
+      status: 401,
+    };
   }
 
-   const nowSec = Math.floor(Date.now() / 1000);
-const loginAtSec = session.user.loginAt
-  ? Math.floor(new Date(session.user.loginAt).getTime() / 1000)
-  : null;
+  if (session.user.role !== "admin") {
+    return {
+      message: "UNAUTHORIZED",
+      success: false,
+      status: 403,
+    };
+  }
 
-  if (!loginAtSec || nowSec - loginAtSec > EXPIRE_SECONDS) {
-  throw new Error("SESSION_EXPIRED");
-}
-
-  const token = session.user.token;
-
-  const result = await fetch(
+  const res = await fetch(
     `${process.env.NEXT_PUBLIC_APP_URL}/api/clients/${clientId}`,
     {
       method: "DELETE",
       headers: {
         "Content-Type": "application/json",
-        Authorization: `Bearer ${token}`,
       },
     }
   );
 
-  if (!result.ok) {
-    throw new Error("FAILED_TO_DELETE_CLIENT");
+  if (!res.ok) {
+    return {
+      message: "Error Deleting Client",
+      success: false,
+      status: res.status,
+    };
   }
 
   revalidatePath("/dashboard/clients");
 
-  return await result.json();
+  return {
+    message: "Client Deleted Successfully",
+    success: true,
+    status: 200,
+  };
 }
