+++ app/[locale]/admin/dashboard/ourTeam/(fetch)/deleteMember.ts
@@ -4,35 +4,50 @@ import { getServerSession } from "next-auth";
 import { revalidatePath } from "next/cache";
 import { authOptions } from "@/app/models/db/authOptions";
 
-const EXPIRE_SECONDS = 15 * 24 * 60 * 60; // 15 days in seconds
-
 export async function deleteMember(memberId: string) {
   const session = await getServerSession(authOptions);
 
   if (!session) {
-    throw new Error("UNAUTHENTICATED");
+    return {
+      message: "UNAUTHENTICATED",
+      success: false,
+      status: 401,
+    };
   }
 
-  const nowSec = Math.floor(Date.now() / 1000);
-const loginAtSec = session.user.loginAt
-  ? Math.floor(new Date(session.user.loginAt).getTime() / 1000)
-  : null;
+  if (session.user.role !== "admin") {
+    return {
+      message: "UNAUTHORIZED",
+      success: false,
+      status: 403,
+    };
+  }
 
-  if (!loginAtSec || nowSec - loginAtSec > EXPIRE_SECONDS) {
-  throw new Error("SESSION_EXPIRED");
-}
-  const token = session.user.token;
+  const res = await fetch(
+    `${process.env.NEXT_PUBLIC_APP_URL}/api/ourTeam/${memberId}`,
+    {
+      method: "DELETE",
+      headers: {
+        "Content-Type": "application/json",
+      },
+    }
+  );
 
-  const result = await fetch(`${process.env.NEXT_PUBLIC_APP_URL}/api/ourTeam/${memberId}`, {
-    method: "DELETE",
-    headers: {
-      "Content-Type": "application/json",
-      Authorization: `Bearer ${token}`,
-    },
-  });
+  if (!res.ok) {
+    return {
+      message: "Error Deleting Member",
+      success: false,
+      status: res.status,
+    };
+  }
 
-  if (!result.ok) throw new Error("FAILED_TO_DELETE_MEMBER");
+  console.log("res: ",res);
+  
+  revalidatePath("/dashboard/ourTeam");
 
-  revalidatePath(`/dashboard/ourTeam`);
-  return await result.json();
+  return {
+    message: "Member Deleted Successfully",
+    success: true,
+    status: 200,
+  };
 }
