+++ app/[locale]/admin/dashboard/users/(fetch)/deleteUser.ts
@@ -1,38 +1,52 @@
 "use server";
+
 import { getServerSession } from "next-auth";
 import { revalidatePath } from "next/cache";
 import { authOptions } from "@/app/models/db/authOptions";
 
-const EXPIRE_SECONDS = 15 * 24 * 60 * 60; // 15 days in seconds
-
-export async function deleteUser(userId:string) {
+export async function deleteUser(userId: string) {
   const session = await getServerSession(authOptions);
-   if (!session) {
-    throw new Error("UNAUTHENTICATED");
-  }
- const nowSec = Math.floor(Date.now() / 1000);
-const loginAtSec = session.user.loginAt
-  ? Math.floor(new Date(session.user.loginAt).getTime() / 1000)
-  : null;
 
-  if (!loginAtSec || nowSec - loginAtSec > EXPIRE_SECONDS) {
-  throw new Error("SESSION_EXPIRED");
-}
-  const token = session?.user.token;
- 
+  if (!session) {
+    return {
+      message: "UNAUTHENTICATED",
+      success: false,
+      status: 401,
+    };
+  }
+
+  if (session.user.role !== "admin") {
+    return {
+      message: "UNAUTHORIZED",
+      success: false,
+      status: 403,
+    };
+  }
+
   const result = await fetch(
     `${process.env.NEXT_PUBLIC_APP_URL}/api/users/editUser/${userId}`,
     {
       method: "DELETE",
       headers: {
         "Content-Type": "application/json",
-        Authorization: `Bearer ${token}`,
+        Authorization: `Bearer ${session.user.token}`,
       },
     }
   );
 
-  if (!result.ok) throw new Error("Failed to delete user");
+  if (!result.ok) {
+    return {
+      message: "Error Deleting User",
+      success: false,
+      status: result.status,
+    };
+  }
 
   revalidatePath(`/dashboard/users`);
-  return await result.json();
+
+  return {
+    message: "User Deleted Successfully",
+    success: true,
+    status: 200,
+  };
 }
